refactor(tv): migrate to new viewmodel
